+++ retrofit/src/main/java/retrofit2/RequestFactory.java
@@ -56,6 +56,7 @@ import retrofit2.http.PUT;
 import retrofit2.http.Part;
 import retrofit2.http.PartMap;
 import retrofit2.http.Path;
+import retrofit2.http.QUERY;
 import retrofit2.http.Query;
 import retrofit2.http.QueryMap;
 import retrofit2.http.QueryName;
@@ -244,6 +245,8 @@ final class RequestFactory {
         parseHttpMethodAndPath("PUT", ((PUT) annotation).value(), true);
       } else if (annotation instanceof OPTIONS) {
         parseHttpMethodAndPath("OPTIONS", ((OPTIONS) annotation).value(), false);
+      } else if (annotation instanceof QUERY) {
+        parseHttpMethodAndPath("QUERY", ((QUERY) annotation).value(), true);
       } else if (annotation instanceof HTTP) {
         HTTP http = (HTTP) annotation;
         parseHttpMethodAndPath(http.method(), http.path(), http.hasBody());